@@ -1,8 +1,6 @@
 import { Component } from '@angular/core';
 import { AuthService } from './core/auth.service';
-import { BranchOfficeService } from './baranch-office/branch-office.service';
 import { BranchOffice } from './baranch-office/branch-office';
-import { SelectedBranchOffice } from './baranch-office/selected-branch-office';
 
 @Component({
   selector: 'app-root',
@@ -17,20 +15,10 @@ export class AppComponent {
   visibilityDrawer = false;
   placementDrawer = 'left';
 
-  constructor(private authService: AuthService, private apiService: BranchOfficeService, public selectedBranchOffice: SelectedBranchOffice) { }
+  constructor(private authService: AuthService) { }
 
   ngOnInit() {
-    this.getBranchOffices();
     this.currentUser = this.authService.getUserName();
-  }
-
-  getBranchOffices() {
-    this.apiService.getBranchOffices().subscribe(data => {
-      this.branchOfficesList = data.sort((a, b) => a.name.localeCompare(b.name));
-      if (this.branchOfficesList.length == 1) {
-        this.selectedBranchOffice.id = this.branchOfficesList[0].id;
-      }
-    });
   }
 
   logout() {
